+++ plugins/modules/zhmc_ldap_server_definition_list.py
@@ -256,8 +256,8 @@ def main():
     # Other exceptions are considered module errors and are handled by Ansible
     # by showing the traceback.
 
-    LOGGER.debug("Module exit (success): changed: %s, result: %r",
-                 changed, result_list)
+    LOGGER.debug("Module exit (success): changed: %s, "
+                 "ldap_server_definitions: %r", changed, result_list)
     module.exit_json(changed=changed, ldap_server_definitions=result_list)