+++ plugins/modules/zhmc_lpar.py
@@ -1256,7 +1256,7 @@ def main():
     # by showing the traceback.
 
     LOGGER.debug(
-        "Module exit (success): changed: %r, cpc: %r", changed, result)
+        "Module exit (success): changed: %r, lpar: %r", changed, result)
     module.exit_json(changed=changed, lpar=result)